requirements.txt
vectorbt==0.24.0
pandas==1.4.3
numpy==1.23.3
statsmodels==0.13.2
yfinance==0.1.74
optuna==3.0.2
tqdm==4.64.0
python-binance==1.0.16
alpaca-trade-api==2.3.0
streamlit==1.13.0